api/requirements.txt
connexion==2.9.0
mysql-connector-python==8.0.26
polyline==1.4.0